Allow empty domain
I've encountered an NTLM-authenticated Exchange server in the wild that allows usernames without the domain name. This works at least in Chrome and curl using NTLM auth. This patch works for me and is the least intrusive I could think of. It also keeps the foot-shooting countermeasures intact, and I'm fine with providing the username to requests-ntlm as "\\foo_user" and not just "foo_user".